improve performance: only write db changes to lix on close
