+++ .changeset/cuddly-rockets-fry.md
@@ -1,0 +1,5 @@
+---
+"@inlang/sdk": patch
+---
+
+improve performance: only write db changes to lix on close
